fix: 统一金额类型为int64(分)并修复BookStatus语义重叠
## 金额类型统一
- 将所有金额字段从 float64(元) 改为 int64(分)
- 涉及模型: Book, Chapter, ChapterPurchase, Wallet, Transaction
- API层自动转换: 用户接口使用元,内部存储使用分

## BookStatus 修复
- 移除 BookStatusPublished 常量
- 统一使用 BookStatusOngoing 表示已发布且正在更新
- 更新所有状态检查逻辑使用 ongoing/completed

## 评分系统统一
- 统一评分范围为 1-5 星
- RatingDistribution 从 map[int]int64 改为 map[string]int64
- 修复 MongoDB BSON 序列化问题

## 测试修复
- 更新测试中的 BookStatus 从 "published" 改为 BookStatusOngoing
- 修复 Mock 对象缺少 BatchUpdateBookStatus 和 UpdateBookStatus 方法
- 归档根目录和 scripts 目录下的临时测试文件

+++ models/bookstore/chapter.go
@@ -15,7 +15,7 @@ type Chapter struct {
 	// Content 字段已移除，使用 ChapterContent 单独存储
 	WordCount int     `bson:"word_count" json:"word_count"`
 	IsFree    bool    `bson:"is_free" json:"is_free"`
-	Price     float64 `bson:"price" json:"price"`
+	Price     int64   `bson:"price" json:"price"` // 价格 (分)
 
 	// 内容引用信息
 	ContentURL     string `bson:"content_url,omitempty" json:"contentUrl,omitempty"`         // 内容获取 URL
@@ -48,8 +48,8 @@ func (c *Chapter) IsPublished() bool {
 	return !c.PublishTime.After(time.Now())
 }
 
-// GetReadingPrice 获取阅读价格
-func (c *Chapter) GetReadingPrice() float64 {
+// GetReadingPrice 获取阅读价格 (分)
+func (c *Chapter) GetReadingPrice() int64 {
 	if c.IsFree {
 		return 0
 	}
